Rename to Spectre.Console.Cli.Extensions.DependencyInjection
