+++ src/Spectre.Console.Cli.Extensions.DependencyInjection.sln
@@ -1,0 +1,34 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Spectre.Console.Cli.Extensions.DependencyInjection", "Spectre.Console.Cli.Extensions.DependencyInjection\Spectre.Console.Cli.Extensions.DependencyInjection.csproj", "{55BC8767-759C-48CD-A927-F5A907E1712F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{55BC8767-759C-48CD-A927-F5A907E1712F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{55BC8767-759C-48CD-A927-F5A907E1712F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{55BC8767-759C-48CD-A927-F5A907E1712F}.Debug|x64.ActiveCfg = Debug|x64
+		{55BC8767-759C-48CD-A927-F5A907E1712F}.Debug|x64.Build.0 = Debug|x64
+		{55BC8767-759C-48CD-A927-F5A907E1712F}.Debug|x86.ActiveCfg = Debug|x86
+		{55BC8767-759C-48CD-A927-F5A907E1712F}.Debug|x86.Build.0 = Debug|x86
+		{55BC8767-759C-48CD-A927-F5A907E1712F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{55BC8767-759C-48CD-A927-F5A907E1712F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{55BC8767-759C-48CD-A927-F5A907E1712F}.Release|x64.ActiveCfg = Release|x64
+		{55BC8767-759C-48CD-A927-F5A907E1712F}.Release|x64.Build.0 = Release|x64
+		{55BC8767-759C-48CD-A927-F5A907E1712F}.Release|x86.ActiveCfg = Release|x86
+		{55BC8767-759C-48CD-A927-F5A907E1712F}.Release|x86.Build.0 = Release|x86
+	EndGlobalSection
+EndGlobal
