+++ src/Spectre.Console.Cli.Extensions.DependencyInjection/DependencyInjectionResolver.cs
@@ -2,7 +2,7 @@ using System;
 using Microsoft.Extensions.DependencyInjection;
 using Spectre.Console.Cli;
 
-namespace Spectre.Cli.Extensions.DependencyInjection
+namespace Spectre.Console.Cli.Extensions.DependencyInjection
 {
     internal class DependencyInjectionResolver : ITypeResolver, IDisposable
     {
