+++ src/Spectre.Console.Cli.Extensions.DependencyInjection/Spectre.Console.Cli.Extensions.DependencyInjection.csproj
@@ -4,14 +4,14 @@
     <TargetFrameworks>netstandard2.0;net7.0</TargetFrameworks>
   </PropertyGroup>
   <PropertyGroup>
-    <PackageId>Spectre.Cli.Extensions.DependencyInjection</PackageId>
-    <Authors>achapman</Authors>
+    <PackageId>Spectre.Console.Cli.Extensions.DependencyInjection</PackageId>
+    <Authors>achapman,devlead,wcomab</Authors>
     <Copyright>© Alistair Chapman 2018</Copyright>
     <PackageTags>spectre;commandline;cli;di;dependency injection;</PackageTags>
-    <RepositoryUrl>https://github.com/agc93/spectre.cli.extensions.dependencyinjection</RepositoryUrl>
+    <RepositoryUrl>https://github.com/wcomab/Spectre.Console.Cli.Extensions.DependencyInjection</RepositoryUrl>
     <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageProjectUrl>https://github.com/agc93/spectre.cli.extensions.dependencyinjection</PackageProjectUrl>
+    <PackageProjectUrl>https://github.com/wcomab/Spectre.Console.Cli.Extensions.DependencyInjection</PackageProjectUrl>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="6.0.1" Condition="'$(TargetFramework)' != 'net7.0'" />
